Example fork
This is an example fork which contains a few JoyGiver classes you can
use in your files.

Each of this adds another type of check:
1) Check if the pawn has a disabled work type whose relevant skills
include the joySkill used.

e.g. if Doctor is disabled, the pawn won't do activities that use the
custom giverClass and have Medicine as joySkill.

2) Check if any of the pawn's disabled work tags disables a skill, and
if that skill is the joySkill used.

e.g. a pawn incapable of Violent can't do activities that give Melee
joySkill if you're using the custom class.

I don't know if this works, it is untested.